+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/organizations/[org_slug]/page.tsx
@@ -1,0 +1,9 @@
+interface OrganizationDashboardPageProps {
+  params: Promise<{ org_slug: string }>;
+}
+export default async function OrganizationDashboardPage({
+  params,
+}: OrganizationDashboardPageProps) {
+  const { org_slug } = await params;
+  return org_slug;
+}
